Rename connected components to follow project conventions
