+++ src/state/components/ConnectedEnergy.js
