+++ src/state/components/ConnectedStage.js
